Initial Zernike fitting module
The initial commit of the Zernike fitting module - the functionality
exists, but the fitting diverges and needs to be investigated.

The zernike.py module also has a class to convert the coefficient files
from the existing CSV files to Zarr. Future commits will include the
functionality to write the fitted Zernike polynomials to either Zarr or
CSV.